fix(checker): repair merge-queue regressions from lib-merged provenance
The merge-group run for #15697 failed on three unit tests and the arch-size
guard; all four are addressed:

- declaration_is_local_to_current_arena now also treats a current-arena node
  the binder bound to the same symbol as local, so a declare-global Array
  augmentation whose NodeIndex collides with a lib declaration keeps its
  members (imported return types no longer report false TS2339).
- Lib-merged TYPE_ALIAS symbols resolve through a dedicated delegation into
  their owning lib context (delegate_lib_merged_symbol_type), running after
  the direct actual-lib shortcuts so canonical name-keyed resolution and its
  materialization ratchets are unaffected. Classes and interfaces keep their
  existing owning paths (intercepting them broke ElementAttributesProperty
  merging under declare-global-in-module: false TS2607).
- The resolve_lazy overlay registration skip is gated on lib-merge
  provenance, so genuine multi-root raw-id collisions keep registering
  their file targets (their cross-file resolution mechanism).
- free_type_params_named added to the common-boundary pub(crate) ratchet
  allowlist; JSX class-signature synthesis and the lib-merged delegation
  moved into their own modules to satisfy the 2000-line size guard.

Verification: jsx_component_attribute_tests 227/227 with real react16/react18
fixtures, conformance_issues_modules + arch_source_scans green, checker unit
failure set a strict subset of main's (32 vs 41, zero new), arch_guard
--size-only passes, clippy clean.
